refactor(monitors): merge resolveNoDataSeverity into computeSeverity
